correção dos exercícios de testes etapa1 a etapa4
